Move threadmanager performance counters out of threadmanager class into separate file
