+++ hpx/include/threadmanager.hpp
@@ -13,6 +13,7 @@
 #include <hpx/runtime/threads/thread_data.hpp>
 #include <hpx/runtime/threads/thread_helpers.hpp>
 #include <hpx/runtime/threads/threadmanager.hpp>
+#include <hpx/runtime/threads/threadmanager_counters.hpp>
 #include <hpx/runtime/threads/topology.hpp>
 
 #endif
